Upgrade the event system and others additions.
Upgrade of the event system in Game, States, StateStack.
The system manage some details to simplify the programming of the events in the game states.
One detail is a trick to fix a bug with the mouse in Microsoft Windows.
Another information about GameObjects : I will probably change the management of the events for them. For this moment, the code of the class is not updated.

For the others additions, the function FrameClock to manage the time of the graphics engine based on the frame rate.
And too, Game and TitleState have some elements to implement a new state. This state is the WorldState and will be the state to play :)
If you download to compile the code, make sure to delete these elements to avoid errors of compilation.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>