Upgrade the title logo.
Upgrade of the title logo in the title state. Now, the logo is animated and I used another homemade program to generate perlin noises.
The code of the animation is not definitive. An animation management will be developped, soon.
And too, trivial modifications in others parts of the code. By example, for this moment, I decided that a frame rate to 30 fps is good :)

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>