Fix some mistakes and Oversights.
Correction of the CMakeLists about the detection of the include directory.
Correction of the perlin noises. Nevertheless, I had no error the day of the publication O_o
Fix a bug with FreeImage which made that the text didn't display. I had fixed that for a long time, and I believed to have published it :o
Deletion of a code to use skybox in the world state, because for this moment the skybox is not ready.
Deletion of the include of Texture2DIdentifier.hpp.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>