Upgrade and fix the vertex buffer objects.
Upgrade the VertexBufferObject class to choose if normales, colors and textures must be initialized.
And too, addition of a parameter to choose the mode to know what kind of primitives to use to render the VBO.
Fixing of the CGDatas in the sending of datas to OpenGL, to delete it in the memory of the client side.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>