Add the skybox and integrate the box.
Addition of the skybox for OpenGL 2.
Integration of the vbo of the box in the world state. You can see it in the center of the universe X)
Changement of the background of the title state to see a rotated skybox. Really, it is the camera which turns :p
The preview background has been deleted.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>