Fix the camera and replace Drimi::Mathematics by GLM.
Cause of no benefits to rewrite my own mathematic module, I replaced Drimi::Mathematics by OpenGL Mathematics.
It's possible that a future version of Drimi::Mathematics will be used as an extension for some things.
Fixing of the camera problem with movement.
Changement in the rotation of the camera in the WorldState.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>