Add OpenGL version switch and integrate shaders.
Integration of the shaders for OpenGL 3.x.
Addition of a system of conditions to use OpenGL 2.x or OpenGL 3.x if the computer is too old or not updated.
Information : For this moment, the shaders are just used to draw in OpenGL 3.x. this is the reason why they are not included in the OpenGL 2.x condition. This system will be better in the future :)

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>