Upgrade the GameObject.
Upgrade of the GameObject to be the atomic element of the game :D
Upgrade of the State to become a game object.
Upgrade of the state stack to take into account the novelties.
Addition of GOs classes for the title state elements.
Upgrade of title state to use the GOs.
Addition of a trivial modification in the game class.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>