Add a part of a system to initialize states and game objects.
Addition of a system to initialize the game states and the game objects to simulate a thread.
When a state is created, after the call by another state, the program was temporary blocked.
This new system serves to solve this problem like as a thread.
The reason I have not used a real thread is the complexity to use OpenGL and SFML in a thread (Risk of conflicts between several graphic objects which don't use the same OpenGL states).
This system is not completed. I again have to upgrade others things like as the textures manager.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>