Upgrade of the states and game objects.
The States inherit now from Game Object.
A new function is added to the Game Objects to prepare the management of the unicode text entering.
Some functions of the States class are transfered to the Game Object class. It's to facilitate the switch between states.
The Event types are transfered to DRIMI.
The error messages in the HandleInput function in the StateStack class are temporary enabled.
The ResizeView functions are now called directly by the StateStack.
Using of a SFML view in the states. It's not really necessary for this moment, but important for the future.

Signed-off-by: TimyusVegetari <ant110283@hotmail.fr>